Add Dec 7, 2024 and OOP idea from Dec 6
Today's wasn't too bad! Part 2 was light compared to the other days 😆.
I think I got lucky with my setup for the part 1, because all I had to
do was make the operators into an argument and add an extra branch for
the concatenation operator. It probably took me 45 minutes in total.

The performance for part 1 was better than I expected, but part 2 is
much slower. It takes around 30 seconds on my machine — I think it must
just be the type conversions and string concatenation.