Add solutions for Dec 9, 2024
This was hard! I really enjoyed the first part because I was able to
improve my algorithm's speed from ~1m3s to ~0.1s. The second puzzle was
tricky. I went through a bunch of iterations and felt super stuck! Then,
I took what I had realized and wrote out some pseudocode and was able to
solve it pretty quickly from there.

It was tricky because I wanted to have cursors looping on both sides,
but I realized that wouldn't work because you have to go back and check
if there's space further on the left.